pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docugenius-cli"
version = "1.0.0"
description = "Convert Word, Excel, PowerPoint, and PDF documents to Markdown format"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "brucevan", email = "brucevanfdm@gmail.com"}
]
keywords = [
    "markdown",
    "converter",
    "docx",
    "pdf",
    "excel",
    "powerpoint",
    "document",
    "docugenius"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "python-pptx>=0.6.23",
    "pdfplumber>=0.10.3",
    "Pillow>=10.0.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/bruc3van/docugenius"
Documentation = "https://github.com/bruc3van/docugenius#readme"
Repository = "https://github.com/bruc3van/docugenius"
Issues = "https://github.com/bruc3van/docugenius/issues"

[project.scripts]
docugenius = "docugenius.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/docugenius"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
